third_party/blink/web_tests/platform/win/external/wpt/html/browsers/windows/clear-window-name.https-expected.txt

This is a testharness.js-based test.
PASS Window.name is not reset when there is an opener around
PASS Window.name is not reset when there is an opener around (cross-origin)
PASS Window.name is not reset at the first navigation away from initial about:blank with noopener
FAIL Window.name is reset at the first cross-origin navigation with noopener assert_equals: expected "" but got "eb079c65-a867-49a6-877f-f814ab0b0da4"
PASS Window.name is not reset at the first navigation away from initial about:blank with window.opener set to null
PASS Window.name is not reset at the same-origin navigation with window.opener set to null
FAIL Window.name is reset at the first first cross-origin navigation with window.opener set to null assert_equals: expected "" but got "3bfa50ca-8ae2-4e17-87d7-40a03a0848db"
PASS Window.name is set by the window
FAIL Window.name is reset at the first cross-origin navigation assert_equals: expected "" but got "7ab69acf-1cc3-477c-85af-3670ff4c79ee"
Harness: the test ran to completion.
